Made Wall Prefab
Something odd happened all the positions in the scene moved so I reset everything to 0,0,0 but not everything went to the middle when I did that... which seems odd that different objects think 0,0,0 is in different places. So not sure what's going on there but for now I created a wall prefab and starting to lay out the environment.