Made Some UI updates and made new items
Okay so I found a food crafting pack. This has lots of food options to craft into meals. Making our job easier for the game. I updated the scripts to use this new items. I'll be updating the chest for level 1 to only contain the items we need for the first crafting. I'll update the bonfire after that