Improved Elemental Staff persistence
The staff list now carries over between scenes. One known glitch is that the EquipDefaultStaff function will add a default staff to the staff list every time you change scenes.  This function may need a check to see if the default staff is already on the list, or else it may need to be removed entirely.